Extract config finder
