+++ lib/config_upgrader.js
@@ -1,11 +1,11 @@
 "use strict";
 
-const Config = require("eslint/lib/config")
+const
+    Config = require("eslint/lib/config")
   , merge = require("eslint/lib/config/config-ops").merge
   , path = require("path")
   , stringify = require("json-stable-stringify")
-  , CONFIG_FILES = require("eslint/lib/config/config-file").CONFIG_FILES
-  , FileFinder = require("eslint/lib/file-finder");
+  ;
 
 var SEVERITIES = {
   0: "off",
@@ -226,34 +226,8 @@ class ConfigUpgrader {
     return [].concat(this._report);
   }
 
-  static configs(analysisFiles) {
-    const dirs = analysisFiles.map(function(fileName){
-      return path.dirname(fileName);
-    });
-    const directories = Array.from(new Set(dirs));
 
-    const localConfigFinder = new FileFinder(CONFIG_FILES, process.cwd());
-
-    const configs = new Set();
-    for (const dir of directories) {
-      const configFiles = localConfigFinder.findAllInDirectoryAndParents(dir);
-      for (const file of configFiles) {
-        configs.add(file);
-      }
-    }
-
-    return Array.from(configs);
-  }
-
-  static upgradeInstructions(configFile, analysisFiles, root) {
-    let configs;
-    if (configFile) {
-      configs = [configFile];
-    }
-    else {
-      configs = this.configs(analysisFiles);
-    }
-
+  static upgradeInstructions(configs, root) {
     const reports = configs.map(function(configFile) {
       let report = [];
