+++ lib/eslint.js
@@ -11,6 +11,7 @@ const checks = require("./checks");
 const validateConfig = require("./validate_config");
 const computeFingerprint = require("./compute_fingerprint");
 const ConfigUpgrader = require("./config_upgrader");
+const findConfigs = require("./config_finder");
 
 const CLIEngine = eslint.CLIEngine;
 const options = { extensions: [".js"], ignore: true, reset: false, useEslintrc: true };
@@ -230,7 +231,14 @@ function run(console, runOptions) {
   if (validateConfig(options.configFile)) {
     console.error("ESLint is running with the " + cli.getConfigForFile(null).parser + " parser.");
 
-    for (const line of ConfigUpgrader.upgradeInstructions(options.configFile, analysisFiles, process.cwd())) {
+
+    let configs = findConfigs(options.configFile, analysisFiles);
+
+    let report = [
+      ConfigUpgrader.upgradeInstructions(configs, process.cwd())
+    ].reduce(function(a, b) { return a.concat([""]).concat(b); });
+
+    for (const line of report) {
       console.error(line);
     }
