+++ test/config_upgrater_test.js
@@ -19,57 +19,9 @@ describe("ConfigUpgrader", function() {
 
 
           let report = ConfigUpgrader
-            .upgradeInstructions(null, [directory + '/file.js'], directory);
+            .upgradeInstructions([directory + '/.eslintrc'], directory);
           expect(report).to.deep.eq([
             ".eslintrc appears to be incompatible with ESLint 3.",
-            "To upgrade it do the following:\n",
-            "* Add .yml or .json to the config file name. Extension-less config file names are deprecated."
-          ]);
-          done();
-        });
-      });
-    });
-
-    it("ignores irrelevant configs", function(done) {
-      temp.mkdir("code ", function(err, directory) {
-        if (err) { throw err; }
-
-        process.chdir(directory);
-
-        const configPath = path.join(directory, ".eslintrc.json");
-        fs.writeFile(configPath, "{}", function(err) {
-          if (err) { throw err; }
-
-          fs.mkdir(path.join(directory, "lib"), function(err) {
-            if (err) { throw err; }
-
-            fs.writeFile(path.join(directory, "lib", ".eslintrc"), "{}", function(err) {
-              if (err) { throw err; }
-
-              let report = ConfigUpgrader
-                .upgradeInstructions(null, [directory + '/file.js'], directory);
-              expect(report).to.deep.eq([]);
-              done();
-            });
-          });
-        });
-      });
-    });
-
-    it("uses specific configs", function(done) {
-      temp.mkdir("code", function(err, directory) {
-        if (err) { throw err; }
-
-        process.chdir(directory);
-
-        const configPath = path.join(directory, "codeclimate-eslint");
-        fs.writeFile(configPath, "{}", function(err) {
-          if (err) { throw err; }
-
-          let report = ConfigUpgrader
-            .upgradeInstructions("codeclimate-eslint", [directory + '/file.js'], directory);
-          expect(report).to.deep.eq([
-            "codeclimate-eslint appears to be incompatible with ESLint 3.",
             "To upgrade it do the following:\n",
             "* Add .yml or .json to the config file name. Extension-less config file names are deprecated."
           ]);
